+++ src/index.jsx
@@ -1,38 +1,16 @@
-/* eslint-disable */
-
-/* the line above disables eslint check for this file (temporarily) todo:delete*/
+/* eslint no-undef: "off" */
+/* reason: 'document' is undefined. */
 
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
-import {
-    createBrowserRouter,
-    RouterProvider,
-} from "react-router-dom";
-
-// import LoginUI from './LoginUI';
-import App from './App';
+// import './index.css';    //not in use
 import reportWebVitals from './reportWebVitals';
+
+import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 
-
-const router = createBrowserRouter([
-    {
-      path: "/",
-      element: <div>Hello world!</div>,
-    },
-  ]);
-
-root.render(
-    <App />
-
-    // <LoginUI/>
-    
-    // <React.StrictMode>
-    //     <RouterProvider router={router} />
-    // </React.StrictMode>
-);
+root.render(<App />);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
